requirements.txt
django
djangorestframework
psycopg2-binary
django-dump-load-utf8==0.0.4
